test(automations): rename strategies module to _runfilter_strategies
The module now contains only run filter strategies; gql_ids and jsonables
move into test_actions.py, their only consumer.
